# api/openapi.yaml
openapi: 3.1.0
info:
  title: TeamTaler API
  version: 0.5.0
  description: |
    Same-origin JSON API for TeamTaler. Monetary values use integer minor units:
    response fields are exact decimal strings, while command inputs are bounded
    JSON integers.
    State-changing authenticated requests require the `X-CSRF-Token` header.
    Product creation, invitation CSV imports, and financial mutations that
    create or reverse bookings, payments, or periods additionally require an
    `Idempotency-Key`.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /groups:
    get:
      tags: [Groups]
      summary: List groups available to the current user
      responses:
        '200':
          description: Active group memberships
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Group' }
    post:
      tags: [Groups]
      summary: Create a group owned by the current user
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, currency]
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
                currency: { type: string, pattern: '^[A-Z]{3}$' }
      responses:
        '201':
          description: Created group
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Group' }
        '422': { $ref: '#/components/responses/Problem' }
  /groups/{groupId}:
    patch:
      tags: [Groups]
      summary: Update administrator-managed group settings
      parameters:
        - $ref: '#/components/parameters/GroupId'
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 120 }
      responses:
        '200':
          description: Persisted normalized group settings
          content:
            application/json:
              schema:
                type: object
                required: [name]
                properties:
                  name: { type: string }
        '403': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: teamtaler_session
  parameters:
    GroupId:
      name: groupId
      in: path
      required: true
      schema: { type: string }
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 1 }
...
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Group:
      type: object
      required: [id, name, currency, membership]
      properties:
        id: { type: string }
        name: { type: string }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        logoUrl: { type: string }
        membership: { $ref: '#/components/schemas/Membership' }
...
    Membership:
      type: object
      required: [id, groupId, userId, email, displayName, status, roles, groupPermissions, categoryGrants]
      properties:
        id: { type: string }
        groupId: { type: string }
        userId: { type: string }
        email: { type: string, format: email }
        displayName: { type: string }
        avatarUrl: { type: string }
        status: { type: string, enum: [ACTIVE, ARCHIVED] }
        roles:
          type: array
          uniqueItems: true
          items: { $ref: '#/components/schemas/Role' }
        groupPermissions:
          type: array
          uniqueItems: true
          items: { $ref: '#/components/schemas/GroupPermission' }
        categoryGrants:
          type: object
          additionalProperties:
            type: array
            uniqueItems: true
            items: { $ref: '#/components/schemas/CategoryPermission' }
    Role:
      type: string
      enum: [ADMIN, FINANCE_MANAGER, CATALOG_MANAGER]
    GroupPermission:
      type: string
      enum: [SELF_RECORD_PAYMENT]
    CategoryPermission:
      type: string
      enum: [ASSIGN_TO_OTHERS, VOID_BOOKINGS]
...
    Problem:
      type: object
      required: [type, title, status, detail, instance]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string }